Home work #2: Refactoring for tests
