+++ app/src/main/java/ru/andrikeev/android/synoptic/application/App.java
@@ -5,6 +5,7 @@ import android.app.Application;
 
 import com.evernote.android.job.JobManager;
 import com.facebook.stetho.Stetho;
+import com.squareup.leakcanary.LeakCanary;
 
 import javax.inject.Inject;
 
@@ -47,6 +48,10 @@ public final class App extends Application implements HasActivityInjector {
         if (BuildConfig.DEBUG) {
             Timber.plant(new Timber.DebugTree());
             Stetho.initializeWithDefaults(this);
+
+            if (!LeakCanary.isInAnalyzerProcess(this)) {
+                LeakCanary.install(this);
+            }
         }
 
         AppInjector.init(this);
