+++ app/src/main/java/ru/andrikeev/android/synoptic/model/repository/WeatherRepositoryImpl.java
@@ -1,0 +1,107 @@
+package ru.andrikeev.android.synoptic.model.repository;
+
+import android.support.annotation.NonNull;
+
+import javax.inject.Inject;
+
+import io.reactivex.Observable;
+import io.reactivex.Single;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.functions.Consumer;
+import io.reactivex.functions.Function;
+import io.reactivex.subjects.PublishSubject;
+import ru.andrikeev.android.synoptic.model.ModelsConverter;
+import ru.andrikeev.android.synoptic.model.data.WeatherModel;
+import ru.andrikeev.android.synoptic.model.network.RemoteService;
+import ru.andrikeev.android.synoptic.model.network.response.WeatherResponse;
+import ru.andrikeev.android.synoptic.model.persistence.CacheService;
+import ru.andrikeev.android.synoptic.model.persistence.Weather;
+import timber.log.Timber;
+
+public class WeatherRepositoryImpl implements WeatherRepository {
+
+    /**
+     * From API documentation: "Do not send requests more than 1 time per 10 minutes
+     * from one device/one API key. Normally the weather is not changing so frequently."
+     */
+    private static final long FETCH_MIN_TIMEOUT = 600_000_000L;
+
+    private static long CITY_STUB_ID = 5601538L; // TODO: change to user choice
+
+    @NonNull
+    private PublishSubject<Resource<WeatherModel>> subject = PublishSubject.create();
+
+    @NonNull
+    private RemoteService remoteService;
+
+    @NonNull
+    private CacheService cacheService;
+
+    @NonNull
+    private ModelsConverter converter;
+
+    @Inject
+    WeatherRepositoryImpl(@NonNull RemoteService remoteService,
+                          @NonNull CacheService cacheService,
+                          @NonNull ModelsConverter converter) {
+        this.remoteService = remoteService;
+        this.cacheService = cacheService;
+        this.converter = converter;
+    }
+
+    private static boolean shouldFetchWeather(@NonNull Weather weather) {
+        return System.currentTimeMillis() - weather.getTimestamp() > FETCH_MIN_TIMEOUT;
+    }
+
+    @NonNull
+    public Observable<Resource<WeatherModel>> loadWeather() {
+        return cacheService.getWeather(CITY_STUB_ID)
+                .onErrorResumeNext(loadRemoteAndSave(CITY_STUB_ID))
+                .map(new Function<Weather, Resource<WeatherModel>>() {
+                    @Override
+                    public Resource<WeatherModel> apply(@NonNull Weather weather) throws Exception {
+                        if (shouldFetchWeather(weather)) {
+                            fetchWeather();
+                            return Resource.fetching(converter.toViewModel(weather));
+                        } else {
+                            return Resource.success(converter.toViewModel(weather));
+                        }
+                    }
+                })
+                .toObservable()
+                .concatWith(subject)
+                .observeOn(AndroidSchedulers.mainThread());
+    }
+
+    @NonNull
+    private Single<Weather> loadRemoteAndSave(long cityId) {
+        return remoteService.getWeather(cityId)
+                .map(new Function<WeatherResponse, Weather>() {
+                    @Override
+                    public Weather apply(@NonNull WeatherResponse weatherResponse) throws Exception {
+                        Timber.d("Weather loaded from api: %s", weatherResponse);
+                        Weather weather = converter.toCacheModel(weatherResponse);
+                        cacheService.cacheWeather(weather);
+                        return weather;
+                    }
+                });
+    }
+
+    public void fetchWeather() {
+        loadRemoteAndSave(CITY_STUB_ID).subscribe(
+                new Consumer<Weather>() {
+                    @Override
+                    public void accept(@NonNull Weather weather) throws Exception {
+                        Timber.d("Weather fetched: %s", weather);
+                        subject.onNext(Resource.success(converter.toViewModel(weather)));
+                    }
+                },
+                new Consumer<Throwable>() {
+                    @Override
+                    public void accept(@NonNull Throwable throwable) throws Exception {
+                        Timber.e(throwable, "Error fetching weather");
+                        subject.onNext(Resource.<WeatherModel>error(throwable));
+                    }
+                });
+    }
+}
